memsystem refactor
Refactor the interface (such as it was) between the processor and the
emulated processor. Took a lot of hackimg and stackoverflow.com advice
to finally settle on something that works.

Added a memory address continuity check after disassembly is finished so
that obvious disassembly guidance issues can be sanity checked.

Combined the exchange instructions into a single data type and
associated operand.

Slimmed down the number of needed ghc extensions to the absolutely
needed set.