memsystem refactor
Refactor the interface (such as it was) between the processor and the
emulated processor. Took a lot of hackimg and stackoverflow.com advice
to finally settle on something that works.

Added a memory address continuity check after disassembly is finished so
that obvious disassembly guidance issues can be sanity checked.

Combined the exchange instructions into a single data type and
associated operand.

Slimmed down the number of needed ghc extensions to the absolutely
needed set.

Promote several of the Z80 pseudo-instructions to DisasemblerTypes
types -- these are commonly used pseudo instructions that ought to
generalize across different processors and systems.

Added an post-decoded-instruction function to the disassemble
type class function in order to filter and adapt to specific situations.

One interesting situation is the 'RST 08' instruction's usage in the
TRS-80 ROM. RST 08 is followed by a byte, which should be automatically
decoded as an ASCII character or ByteRange pseudo-operation.

Use Data.Text instead of Data.ByteString
Compute MD5 hash on the ROM, prepend to the disassembly output.