refactor disassembler pseudo-instructions
Promote several of the Z80 pseudo-instructions to DisasemblerTypes
types -- these are commonly used pseudo instructions that ought to
generalize across different processors and systems.

Added an post-decoded-instruction function to the disassemble
type class function in order to filter and adapt to specific situations.

One interesting situation is the 'RST 08' instruction's usage in the
TRS-80 ROM. RST 08 is followed by a byte, which should be automatically
decoded as an ASCII character or ByteRange pseudo-operation.