checkpoint_130102
Found a subtle bug in the instruction decoder: the decoder dispatch
function was checking the current opcode, opc, whereas the dispatch
function should have an opcode as a parameter because the program
counter is incremented when dealing with index-prefixed instructions.

Also found that data constructors HLIndirect, IXIndirect and IYindirect
are not used, removed them.

Refactored the Misosys EDAS-compatible assembler source, now that some
clarity is setting in. Now has a test harness so that some validation of
the assembler can happen using Test.HUnit. And an initial test of
equates is already working.

Add a trailing newline to source that doesn't have one already. This
allows the parser to succeed without extra stuff in the test harness,
e.g., an extra Data.Text..empty so that Data.Text.intercalate will
automagically add one to the last element of its list.