rework devices
Hopefully, this is a cleaner design. Created a boxed type, Device, that
provides the appropriate amount of polymorphism so that MemorySystem can
manage collections of memory-mapped devices. The "reference" to the
mmap-ed device is an Int that is a key into the mmap-ed device container
held by the MemorySystem.

Devices, themselves, have embedded State transformer functions, in
addition to encapsulating device state.