round out test video device
Add proper read/write functions for the test video device. Note that the
test video device has a specific internal Word16 address and Word8 word
types, which matches the test system's memory system. Otherwise, there
would need to be an address and word type-specific DeviceIO instance
that does the appropriate address/word type conversions (via
fromIntegral, most likely.)

Minor change to mRead: Pass the offset from the start of the memory
region to the device instead of the absolute address. This cuts down on
code and things to remember in the actual devices (they don't have to
store the base address in order to calculate offsets, which is what
they'd normally have to do.)